Move isExecutable to ExpressionStatement
In the future, we will want to check for executability inside
programs as well as dialogue states.
